+++ Adj/Algebra.hs
@@ -5,3 +5,4 @@ import Adj.Algebra.Category as Exports
 import Adj.Algebra.Functor as Exports
 import Adj.Algebra.Morphism as Exports
 import Adj.Algebra.Product as Exports
+import Adj.Algebra.Sum as Exports