+++ Adj/Algebra/Sum.hs
@@ -1,0 +1,5 @@
+module Adj.Algebra.Sum where
+
+infixr 7 :+:
+
+data (:+:) left right = Option left | Adoption right