+++ CHANGELOG.md
@@ -13,3 +13,4 @@
 * Define `(-/->)` type alias
 * Define `(<-\-)` type alias
 * Define `Adj.Algebra.Product` module
+* Define `Adj.Algebra.Sum` module